requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.8.0
torchvision==0.23.0
torchdiffeq==0.2.5

scikit-learn==1.8.0
pyyaml==6.0.3
matplotlib==3.10.8